Admin can now update or delete users posts

+++ back/controllers/postsController.js
@@ -1,5 +1,6 @@
 const fs = require("fs")
 const Post = require("../models/postModel")
+const Like = require("../models/likeModel")
 
 // Demande à la DB de renvoyer tous les documents de la collection Post.
 exports.getAllPosts = (req, res) => {
@@ -30,7 +31,6 @@ exports.getPost = (req, res) => {
 // Crée un nouveau post avec l'id de l'utilisateur, les informations qu'il saisit sur la page d'envoi, le chemin d'accès à l'image reçue, et initialise le nombre de likes à 0. Array d'utilisateurs ayant liké est donc de facto vide aussi.
 exports.createPost = async (req, res) => {
   let imageUrl
-  // const postObject = req.body
   if (req.file) {
     imageUrl = `${req.protocol}://${req.get("host")}/images/${
       req.file.filename
@@ -39,7 +39,7 @@ exports.createPost = async (req, res) => {
     imageUrl = ""
   }
   await Post.create({
-    userId: req.body.userId,
+    userId: req.user.userId,
     text_content: req.body.text_content,
     likes: 0,
     image_url: imageUrl,
@@ -62,40 +62,46 @@ exports.updatePost = (req, res, next) => {
     where: { id: req.params.id },
   })
     .then((post) => {
-      if (req.file) {
-        const imgPath = post.image_url.replace(
-          "http://127.0.0.1:3000",
-          "."
-        )
-        fs.unlink(imgPath, (err) => {
-          if (err) {
-            console.error(err)
-          } else {
-            console.log("Image supprimée")
+      if (post.userId !== req.body.userId) {
+        if (req.user.role == 0) {
+          return next(res.status(401))
+        } else {
+          if (req.file) {
+            const imgPath = post.image_url.replace(
+              "http://127.0.0.1:3000",
+              "."
+            )
+            fs.unlink(imgPath, (err) => {
+              if (err) {
+                console.error(err)
+              } else {
+                console.log("Image supprimée")
+              }
+            })
           }
-        })
+          const postObject = req.file
+            ? {
+                userId: post.userId,
+                text_content: req.body.text_content,
+                image_url: `${req.protocol}://${req.get(
+                  "host"
+                )}/images/${req.file.filename}`,
+              }
+            : {
+                userId: post.userId,
+                text_content: req.body.text_content,
+                image_url: post.imageUrl,
+              }
+          Post.upsert({
+            id: req.params.id,
+            ...postObject,
+          })
+            .then(() =>
+              res.status(200).json({ message: "Post mis à jour." })
+            )
+            .catch((error) => res.status(400).json({ error }))
+        }
       }
-      const postObject = req.file
-        ? {
-            userId: post.userId,
-            text_content: req.body.text_content,
-            image_url: `${req.protocol}://${req.get("host")}/images/${
-              req.file.filename
-            }`,
-          }
-        : {
-            userId: post.userId,
-            text_content: req.body.text_content,
-            image_url: post.imageUrl,
-          }
-      Post.upsert({
-        id: req.params.id,
-        ...postObject,
-      })
-        .then(() =>
-          res.status(200).json({ message: "Post mis à jour." })
-        )
-        .catch((error) => res.status(400).json({ error }))
     })
     .catch((error) => res.status(400).json({ error }))
 }
@@ -107,34 +113,36 @@ exports.deletePost = (req, res, next) => {
   })
     .then((post) => {
       if (post.userId !== req.body.userId) {
-        return next(res.status(401))
-      } else {
-        Post.destroy({ where: { id: req.params.id } })
-          .then((post) => {
-            if (post.imageUrl !== undefined) {
-              const imgPath = post.imageUrl.replace(
-                "http://localhost:3000",
-                "."
-              )
-              fs.unlink(imgPath, (err) => {
-                if (err) {
-                  console.error(err)
-                } else {
-                  console.log("Image supprimée")
-                }
-              })
-            }
-          })
-          .then(() => {
-            res
-              .status(200)
-              .json({ message: "Le post a bien été supprimé." })
-          })
-          .catch((error) => {
-            res.status(400).json({
-              error: error,
+        if (req.user.role == 0) {
+          return next(res.status(401))
+        } else {
+          Post.destroy({ where: { id: req.params.id } })
+            .then((post) => {
+              if (post.imageUrl !== undefined) {
+                const imgPath = post.imageUrl.replace(
+                  "http://localhost:3000",
+                  "."
+                )
+                fs.unlink(imgPath, (err) => {
+                  if (err) {
+                    console.error(err)
+                  } else {
+                    console.log("Image supprimée")
+                  }
+                })
+              }
             })
-          })
+            .then(() => {
+              res
+                .status(200)
+                .json({ message: "Le post a bien été supprimé." })
+            })
+            .catch((error) => {
+              res.status(400).json({
+                error: error,
+              })
+            })
+        }
       }
     })
     .catch((error) => {
@@ -146,16 +154,20 @@ exports.deletePost = (req, res, next) => {
 
 // Trouve le post correspondant à l'id de la page, puis analyse le corps de la requête pour savoir si l'utilisation a liké ou annulé un like. Si l'utilisateur like, son id est enregistré dans un array et le compteur associé augmente de 1. S'il s'agit d'une annulation de like, supprime l'ID de l'utilisateur de l'array et diminue le nombre de like en question de 1.
 exports.likePost = (req, res) => {
-  Post.findByPk(req.params.id)
-    .then((post) => {
-      if (req.body.like === 1) {
-        return post.update({
-          $push: { usersLiked: req.body.userId },
-          $inc: { likes: 1 },
-        })
-      } else if (req.body.like === 0) {
+  Post.findOne({
+    where: { id: req.params.id },
+  })
+  .then((post) => {
+    // console.log(post)
+    console.log(req.body)
+    if (req.body.like === 1) {
+      return post.upsert({
+        // $push: { usersLiked: req.body.userId },
+        $inc: { likes: 1 },
+      })
+    } else if (req.body.like === 0) {
         if (post.usersLiked.includes(req.body.userId)) {
-          return post.update({
+          return post.upsert({
             $pull: {
               usersLiked: req.body.userId,
             },
